refactor(proprietary): move scrapers/ -> platforms/ (one subpackage per platform)
