+++ surfsense_backend/app/proprietary/platforms/__init__.py
@@ -1,7 +1,6 @@
-"""Platform-specific crawl/extraction actors (non-Apache-2; see ../LICENSE).
+"""Platform-native crawl/extraction actors (non-Apache-2; see ../LICENSE).
 
-Scaffolded for future work (Phase 8 — Platform Actors). Each platform (e.g.
-maps, professional networks) will get its own subpackage that reuses the shared
-fetch strategies in ``app.proprietary.web_crawler`` under a platform-specific
-structured extractor. Empty for now by design.
+One subpackage per platform (e.g. ``youtube``), each a structured extractor for
+that platform's data. Actors may reuse the shared fetch strategies in
+``app.proprietary.web_crawler`` but expose their own platform-specific I/O.
 """
