--- surfsense_backend/app/proprietary/scrapers/__init__.py
@@ -1,1 +1,0 @@
-"""Platform-native scrapers (non-connector data extraction tools)."""
