+++ surfsense_backend/tests/unit/platforms/__init__.py
@@ -1,0 +1,1 @@
+"""Platform-native scraper packages live here (one subpackage per platform)."""
